refactor(bin): begin layout of Pepp IDE
