--- bin/demo/layout/fig-view_en_US.ts
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE TS>
-<TS version="2.1" language="en_US"></TS>
