--- bin/demo/layout/main.cpp
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2023 J. Stanley Warford, Matthew McRaven
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#include <QDirIterator>
-#include <QGuiApplication>
-#include <QQmlApplicationEngine>
-#include "shared.hpp"
-#include <QLocale>
-#include <QQuickItem>
-#include <QTranslator>
-#include <qvariant.h>
-
-//#include <QDirIterator>
-#include <QQmlContext>
-
-
-QSharedPointer<gui_globals> initializeLayout(QQmlApplicationEngine &engine) {
-  return nullptr;
-}
-
