+++ bin/gui/about/AboutDialog.qml
