+++ bin/gui/components/BinaryEdit.qml
