+++ bin/gui/components/DisplaySetting.qml
