+++ bin/gui/components/HexEdit.qml
