+++ bin/gui/components/NameValueEdit.qml
