+++ bin/gui/components/RegisterBinEdit.qml
