+++ bin/gui/components/RegisterHexEdit.qml
