+++ bin/gui/components/RowNumber.qml
