+++ bin/gui/cpu/registermodel.cpp
