+++ bin/gui/cpu/statusbitmodel.cpp
