+++ bin/gui/cpu/statusbitmodel.h
