+++ bin/gui/theme/Theme.qml
