+++ bin/gui/theme/qmldir
